+++ src/containerapp/azext_containerapp/_clients.py
@@ -26,7 +26,7 @@ from knack.log import get_logger
 
 logger = get_logger(__name__)
 
-PREVIEW_API_VERSION = "2024-02-02-preview"
+PREVIEW_API_VERSION = "2024-08-02-preview"
 AUG_PREVIEW_API_VERSION = "2024-08-02-preview"
 POLLING_TIMEOUT = 1500  # how many seconds before exiting
 POLLING_SECONDS = 2  # how many seconds between requests