+++ src/containerapp/azext_containerapp/_params.py
@@ -442,9 +442,10 @@ def load_arguments(self, _):
         c.argument('all', help="The flag to indicate all logger settings.", action="store_true")
 
     with self.argument_context('containerapp debug') as c:
-        c.argument('container', help="The container name that the ephemeral container will target to.")
+        c.argument('container',
+                   help="The container name that the debug console will connect to. Defaults to the first container of first replica.")
         c.argument('replica',
-                   help="The name of the replica. List replicas with 'az containerapp replica list'. A replica may not exist if there is no traffic to your app.")
+                   help="The name of the replica. List replicas with 'az containerapp replica list'. A replica may be not found when it's scaled to zero if there is no traffic to your app. Defaults to the first replica of 'az containerapp replica list'.")
         c.argument('revision',
                    help="The name of the container app revision. Defaults to the latest revision.")
         c.argument('name', name_type, id_part=None, help="The name of the Containerapp.")