+++ src/containerapp/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 upcoming
 ++++++
-* 'az containerapp debug': Open an SSH-like interactive shell within a container app ephemeral container.
+* 'az containerapp debug': Open an SSH-like interactive shell within a container app debug console.
 
 1.0.0b1
 ++++++